+++ Repository/DTO/ResponseDTO/CarRentalRate/CarRentalRateView.cs
@@ -9,14 +9,14 @@ namespace Repository.DTO.ResponseDTO.CarRentalRate
 {
     public class CarRentalRateView
     {
-        public double? DailyRate { get; set; }
-        public double? HourlyRate { get; set; }
+        public decimal? DailyRate { get; set; }
+        public decimal? HourlyRate { get; set; }
 
         public double? WeeklyDiscount { get; set; }
         public double? MonthlyDiscount { get; set; }
 
         public int? MaxDistancePerDay { get; set; }
-        public double? OvertravelRatePerKm { get; set; }
+        public decimal? OvertravelRatePerKm { get; set; }
 
         public string Status { get; set; }
